requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==4.2
django-allauth==0.63.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.1
PyJWT==2.8.0
sqlparse==0.5.1
typing-extensions==4.12.2
whitenoise==6.7.0
